Rename test project
