+++ source/App/App.sln
@@ -42,11 +42,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "example", "example", "{856E
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExampleHost.FunctionApp01", "source\ExampleHost.FunctionApp01\ExampleHost.FunctionApp01.csproj", "{FB0C3D18-4366-421B-B5C6-56F0A7D7BA16}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExampleHost.Tests", "source\ExampleHost.Tests\ExampleHost.Tests.csproj", "{DFB36594-59B4-4ABB-A43D-BD78A648930F}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExampleHost.FunctionApp02", "source\ExampleHost.FunctionApp02\ExampleHost.FunctionApp02.csproj", "{F1E201ED-4DA7-49A0-A132-F0A48D109D94}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleHost.WebApi01", "source\ExampleHost.WebApi01\ExampleHost.WebApi01.csproj", "{A182F9DE-6C59-49C1-9EB8-E9DD12C522FE}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExampleHost.FunctionApp.Tests", "source\ExampleHost.FunctionApp.Tests\ExampleHost.FunctionApp.Tests.csproj", "{530AAF08-2CEC-47FE-B717-A7A07C7E3195}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -94,10 +94,6 @@ Global
 		{FB0C3D18-4366-421B-B5C6-56F0A7D7BA16}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FB0C3D18-4366-421B-B5C6-56F0A7D7BA16}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FB0C3D18-4366-421B-B5C6-56F0A7D7BA16}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DFB36594-59B4-4ABB-A43D-BD78A648930F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DFB36594-59B4-4ABB-A43D-BD78A648930F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DFB36594-59B4-4ABB-A43D-BD78A648930F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DFB36594-59B4-4ABB-A43D-BD78A648930F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F1E201ED-4DA7-49A0-A132-F0A48D109D94}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F1E201ED-4DA7-49A0-A132-F0A48D109D94}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F1E201ED-4DA7-49A0-A132-F0A48D109D94}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -106,6 +102,10 @@ Global
 		{A182F9DE-6C59-49C1-9EB8-E9DD12C522FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A182F9DE-6C59-49C1-9EB8-E9DD12C522FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A182F9DE-6C59-49C1-9EB8-E9DD12C522FE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{530AAF08-2CEC-47FE-B717-A7A07C7E3195}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{530AAF08-2CEC-47FE-B717-A7A07C7E3195}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{530AAF08-2CEC-47FE-B717-A7A07C7E3195}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{530AAF08-2CEC-47FE-B717-A7A07C7E3195}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -122,9 +122,9 @@ Global
 		{0C3952A7-BF54-4467-A70F-41EC5F3C3A0D} = {6B41AD50-3A7B-41DE-8FF7-E74704460417}
 		{432DA871-1514-423E-B0EB-7F5873CB1E45} = {6B41AD50-3A7B-41DE-8FF7-E74704460417}
 		{FB0C3D18-4366-421B-B5C6-56F0A7D7BA16} = {856E0E44-CB7A-4A3F-8243-3B5601B0A553}
-		{DFB36594-59B4-4ABB-A43D-BD78A648930F} = {856E0E44-CB7A-4A3F-8243-3B5601B0A553}
 		{F1E201ED-4DA7-49A0-A132-F0A48D109D94} = {856E0E44-CB7A-4A3F-8243-3B5601B0A553}
 		{A182F9DE-6C59-49C1-9EB8-E9DD12C522FE} = {856E0E44-CB7A-4A3F-8243-3B5601B0A553}
+		{530AAF08-2CEC-47FE-B717-A7A07C7E3195} = {856E0E44-CB7A-4A3F-8243-3B5601B0A553}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {97C97DEA-A677-4CDB-919D-DECF30991C7B}
