+++ source/App/source/ExampleHost.FunctionApp.Tests/integrationtest.local.settings.sample.json
